sam2-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
numpy==1.26.3
# NOTE: PyTorch and MedSAM2 are installed separately per CLAUDE.md
# Do NOT add torch here — it must be installed with the CUDA index URL
